fix hive
